refactor: moving post API
