+++ routes/post.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+
+const {
+  getPostByView,
+  getPostByLike,
+} = require("../controllers/postController");
+
+router.route("/api/getPostByView").get(getPostByView);
+router.route("/api/getPostByLike").get(getPostByLike);
